Move RNE-specific derivative tests into `derivative_test.py` .
